feat(#32): Update post model
- Make status of type `short` so it's easily mapped to postgresql smallint
- Create custom get/setStatus so the short value can be mapped to/from the PostStatus enum
- Update PostStatus to include a specifically defined value. Add utility methods to map to and from a short value (from the defined ones)
- Update PostRepository to receive a `short` value to be executed in the jql for getPosts
- Update UTs

+++ persistence/src/test/java/org/wcdevs/blog/core/persistence/post/PostTest.java
@@ -130,7 +130,7 @@ class PostTest {
                    .updatedOn(updatedOn)
                    .publishedBy(publishedBy)
                    .updatedBy(updatedBy)
-                   .status(PostStatus.PUBLISHED)
+                   .status(PostStatus.PUBLISHED.shortValue())
                    .build();
 
     assertEquals(title, post.getTitle());
